Restructure log and exception exports.
