+++ packages/powersync/lib/powersync.dart
@@ -3,11 +3,12 @@
 /// Use [PowerSyncDatabase] to open a database.
 library;
 
-export 'src/powersync_database.dart';
-export 'src/schema.dart';
 export 'src/connector.dart';
 export 'src/crud.dart';
+export 'src/exceptions.dart';
+export 'src/log.dart';
+export 'src/open_factory.dart';
+export 'src/powersync_database.dart';
+export 'src/schema.dart';
 export 'src/sync_status.dart';
 export 'src/uuid.dart';
-export 'src/open_factory.dart';
-export 'src/log.dart' show attachedLogger, autoLogger, debugLogger;
